+++ src/combinators/map.test.js
@@ -12,19 +12,16 @@ describe('#map', () => {
     completeSpy = jest.fn()
   })
 
-  it('maps a function over the signal values', done => {
+  it('maps a function over the signal values', () => {
     const s = Signal.fromArray(range(1, 3))
 
     map(inc)(s).subscribe(valueSpy, errorSpy, completeSpy)
 
-    setTimeout(() => {
-      range(2, 3).forEach((n, index) => {
-        expect(valueSpy.mock.calls[index][0]).toBe(n)
-      }, this)
+    range(2, 3).forEach((n, index) => {
+      expect(valueSpy.mock.calls[index][0]).toBe(n)
+    }, this)
 
-      expect(completeSpy).toHaveBeenCalled()
-      done()
-    }, 0)
+    expect(completeSpy).toHaveBeenCalled()
   })
 
   it('emits an error if the parent signal emits an error', () => {