+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Add `Signal#cycle` method
 * Add `Signal#sequential` method
 * Deprecate `Signal.sequential` method (use `Signal.periodic(1000).sequential([1, 2, 3])` instead)
+* Remove `setTimeout` in `Signal.fromArray` method
+* Remove `setTimeout` in `scan` function
 
 ## 1.3.0 (2019-01-06)